refactor(cli): name the MCP launch-command transform after what it does
`tools/domain/mcp-exclusion.ts` holds only `transformFor`, the transform of an
MCP server's launch command for the target platform. The exclusion it was
named for was deleted with #816, so the name pointed at a feature that no
longer exists. The module becomes `mcp-launch-command.ts`, with its test file,
its one import and its entry in the context-boundary public-module list.
No behaviour change.

Fixes #824

AIDD-Session-Id: 4acc9a1c-19bc-4468-b8b6-e86644bcba60
