+++ cli/src/contexts/tools/domain/mcp-launch-command.ts
