+++ cli/tests/architecture/context-boundary.arch.test.ts
@@ -15,7 +15,7 @@ const PUBLIC_MODULES: Readonly<Record<string, readonly string[]>> = {
     "src/contexts/tools/domain/build-contract.ts",
     // co-owned configuration (settings.json, .mcp.json et al.)
     "src/contexts/tools/domain/capabilities/mcp-capability.ts",
-    "src/contexts/tools/domain/mcp-exclusion.ts",
+    "src/contexts/tools/domain/mcp-launch-command.ts",
     "src/contexts/tools/domain/capabilities/settings-capability.ts",
     "src/contexts/tools/domain/capabilities/hooks-capability.ts",
     "src/contexts/tools/domain/capabilities/config-refs.ts",
